requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2019.9.11
cffi==1.13.1
Click==7.0
entrypoints==0.3
flake8==3.7.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.2.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
wincertstore==0.2
WTForms==2.2.1
